+++ adapter-vungle/CloudXVungleAdapter.xcframework/ios-arm64/CloudXVungleAdapter.framework/Headers/CLXVungleMetadataProvider.h
@@ -1,0 +1,10 @@
+#import <Foundation/Foundation.h>
+#import <CloudXCore/CLXAdapterMetadataProvider.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface CLXVungleMetadataProvider : NSObject <CLXAdapterMetadataProvider>
++ (instancetype)createInstance;
+@end
+
+NS_ASSUME_NONNULL_END